Rename rpc trait file
