+++ core/src/traits/mod.rs
@@ -1,1 +1,1 @@
-pub mod verifier;
+pub mod rpc;
